ptb/requirements.txt
python-telegram-bot==20.4
requests==2.28.2
pymongo